Speed up CRC computation
Rather than reverting every byte, it is faster to apply the reverted
polynomial algorithm[1]. With that change, I able to generate .rps file
14x time faster (0.1s vs 1.4s).

Note in my previous implementation, I tries to compute the CRC table
offline. It appears it is faster to compute the table on runtime rather
than loading the Python array.

I have check the resulting binaries are exactly the same.

[1]: https://zlib.net/crc_v3.txt (section 11.)